Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/getBoundingClientRect-empty-inline-expected.txt


FAIL getBoundingClientRect-empty-inline assert_equals: y expected 92 but got 0
